Add test step references in documentation
